added conversion ratio

+++ Core/Inc/bq79656.h
@@ -327,7 +327,7 @@ void BQ_Main();
 void BQ_SetFaultState();
 void BQ_Init();
 int32_t BQ_GetCurrent();
-int16_t BQ_GetVoltage(int cell);
+float BQ_GetVoltage(int cell);
 
 void BQ_AutoAddressing();
 void BQ_ReadVoltages();
